refactor: share sleep helper and inter-step delay constant via utils/timing
flow-execute and run-sequence both want the same 100ms inter-step delay;
extract sleep + DEFAULT_INTER_STEP_DELAY_MS into a dedicated timing util
so changing the default is a one-line edit. Moves sleep out of
gesture-utils (where it didn't belong) and updates gesture-pinch /
gesture-rotate to follow the new path.
