+++ packages/tool-server/src/tools/run-sequence/index.ts
@@ -2,8 +2,7 @@ import { z } from "zod";
 import type { Registry, ToolCapability, ToolDefinition } from "@argent/registry";
 import { simulatorServerRef } from "../../blueprints/simulator-server";
 import { resolveDevice } from "../../utils/device-info";
-
-const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
+import { sleep, DEFAULT_INTER_STEP_DELAY_MS } from "../../utils/timing";
 
 const ALLOWED_TOOLS = new Set([
   "gesture-tap",
@@ -36,7 +35,9 @@ const zodSchema = z.object({
         delayMs: z
           .number()
           .optional()
-          .describe("Wait time in ms after this step before the next (default 100)"),
+          .describe(
+            `Wait time in ms after this step before the next (default ${DEFAULT_INTER_STEP_DELAY_MS})`
+          ),
       })
     )
     .min(1)
@@ -135,7 +136,7 @@ Stops on the first error and returns partial results.`,
           break;
         }
 
-        const delay = step.delayMs ?? 100;
+        const delay = step.delayMs ?? DEFAULT_INTER_STEP_DELAY_MS;
         if (delay > 0) await sleep(delay);
       }
