+++ packages/tool-server/src/utils/gesture-utils.ts
@@ -1,8 +1,6 @@
 import type { SimulatorServerApi } from "../blueprints/simulator-server";
 import { sendCommand } from "./simulator-client";
 
-export const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
-
 interface TouchPoint {
   x: number;
   y: number;
